Upgraded UI to play using clickable images
I have greatly improved the UI for the game. The UI consists of three
images of rock paper and scissors which can be clicked to play.
Additionally, the selection for the program automatically using a time
interval. When the player chooses an image to click, the interval stops
and displays a random pick by the program. Ultimately, the scores are
updated and a winner is selected when either the player or the program
reached 5 points. A message pops up and the player can try again.